Updated code to support utils helper file.
