+++ chrome-extension-widify/manifest.json
@@ -38,6 +38,7 @@
     }
   ],
   "background": {
-    "service_worker": "background.js"
-  }
+    "service_worker": "background.js",
+    "type": "module"
+}
 }
