+++ chrome-extension-widify/options.html
@@ -6,7 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>WIDIFY</title>
     <link rel="stylesheet" href="options.css" />
-    <script type="text/javascript" src="options.js" defer></script>
+    <script type="module" src="./utils.js"></script>
+    <script type="module" src="./options.js"></script>
   </head>
 
   <body>
